1.0.0

+++ neiko-modules/neiko-system/neiko-system-api/src/main/java/com/schilings/neiko/system/controller/SysUserController.java
@@ -76,7 +76,7 @@ public class SysUserController {
 	@GetMapping("/{userId}")
 	@PreAuthorize(value = "hasAuthority('system:user:read')")
 	@Operation(summary = "获取指定用户的基本信息")
-	public R<SysUserInfo> getSysUserInfo(@PathVariable("userId") Integer userId) {
+	public R<SysUserInfo> getSysUserInfo(@PathVariable("userId") Long userId) {
 		SysUser sysUser = sysUserService.getById(userId);
 		if (sysUser == null) {
 			return R.ok();
